+++ CHANGELOG.md
@@ -30,6 +30,8 @@
   e.g. `vim -S Session1.vim`.
 - Enable restoring command arguments for inline strategies with `*` character.
 - Kill session "0" if it wasn't restored.
+- Add `@resurrect-delete-backup-after` option to specify how many days of
+  backups to keep - default is 30.
 
 ### v2.4.0, 2015-02-23
 - add "tmux-test"